+++ packages/events/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./eventBus.js";
 export * from "./inMemoryEventBus.js";
+export * from "./durableEventBus.js";